Analysis using only A and B genes (#16)
In this PR I have created a new nb: 1_main_Pa_sim_enhance_AtoB_simplified_input.ipynb

This will perform the analysis (#3) mentioned in comments from @danich1 in PR #15

Here we remove all non-A,B genes and redo the same analysis where we manually shift the input gene expression data and encode and decode it using Tybalt. Then we look to see the trends that are learned.

Still need to work on the interpretation of these results but happy to brainstorm. One interesting pattern is that the plot for A vs A' is inverted which we wouldn't expect.

* rename files with index

* add script with simplified input data and update scripts

* add intermediate files